requirements.txt
Django~=2.1.5
Django-crispy-forms==1.7.2
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.1
django-taggit==0.22.2
pillow==5.4.1
django-reservation
boto3
Braintree==3.45.0
django-storages==1.7.1
